# backend/ehr-service/openapi.yaml
openapi: 3.1.0
info:
  title: EHOS EHR Service
  version: 0.1.0
paths:
  /api/v1/patients/{patient_id}/encounters:
    post:
      tags:
      - ehr
      summary: Create Encounter
      operationId: create_encounter_api_v1_patients__patient_id__encounters_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncounterIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - ehr
      summary: List Encounters
      operationId: list_encounters_api_v1_patients__patient_id__encounters_get
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/encounters/{encounter_id}/close:
    post:
      tags:
      - ehr
      summary: Close Encounter
      operationId: close_encounter_api_v1_patients__patient_id__encounters__encounter_id__close_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: encounter_id
        in: path
        required: true
        schema:
          type: string
          title: Encounter Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/notes:
    post:
      tags:
      - ehr
      summary: Create Note
      operationId: create_note_api_v1_patients__patient_id__notes_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - ehr
      summary: List Notes
      operationId: list_notes_api_v1_patients__patient_id__notes_get
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: note_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Note Type
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/notes/{note_id}:
    get:
      tags:
      - ehr
      summary: Get Note
      operationId: get_note_api_v1_patients__patient_id__notes__note_id__get
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - ehr
      summary: Update Note
      operationId: update_note_api_v1_patients__patient_id__notes__note_id__patch
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/notes/{note_id}/amend:
    post:
      tags:
      - ehr
      summary: Amend Note
      operationId: amend_note_api_v1_patients__patient_id__notes__note_id__amend_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmendmentIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/notes/{note_id}/sign:
    post:
      tags:
      - ehr
      summary: Sign Note
      operationId: sign_note_api_v1_patients__patient_id__notes__note_id__sign_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      - name: signed_by
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Signed By
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/notes/{note_id}/versions:
    get:
      tags:
      - ehr
      summary: List Versions
      operationId: list_versions_api_v1_patients__patient_id__notes__note_id__versions_get
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/notes/{note_id}/amendments:
    get:
      tags:
      - ehr
      summary: List Amendments
      operationId: list_amendments_api_v1_patients__patient_id__notes__note_id__amendments_get
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/soap:
    post:
      tags:
      - ehr
      summary: Create Soap
      operationId: create_soap_api_v1_patients__patient_id__soap_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SOAPNoteIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/progress-notes:
    post:
      tags:
      - ehr
      summary: Create Progress
      operationId: create_progress_api_v1_patients__patient_id__progress_notes_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressNoteIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/discharge-summary:
    post:
      tags:
      - ehr
      summary: Create Discharge
      operationId: create_discharge_api_v1_patients__patient_id__discharge_summary_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DischargeSummaryIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/vitals:
    post:
      tags:
      - ehr
      summary: Record Vitals
      operationId: record_vitals_api_v1_patients__patient_id__vitals_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - ehr
      summary: List Vitals
      operationId: list_vitals_api_v1_patients__patient_id__vitals_get
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: vital_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Vital Type
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/diagnoses:
    post:
      tags:
      - ehr
      summary: Add Diagnosis
      operationId: add_diagnosis_api_v1_patients__patient_id__diagnoses_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagnosisIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - ehr
      summary: List Diagnoses
      operationId: list_diagnoses_api_v1_patients__patient_id__diagnoses_get
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}/diagnoses/{diagnosis_id}/resolve:
    post:
      tags:
      - ehr
      summary: Resolve Diagnosis
      operationId: resolve_diagnosis_api_v1_patients__patient_id__diagnoses__diagnosis_id__resolve_post
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          title: Patient Id
      - name: diagnosis_id
        in: path
        required: true
        schema:
          type: string
          title: Diagnosis Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagnosisResolveIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AmendmentIn:
      properties:
        amendment:
          type: string
          maxLength: 20000
          minLength: 1
          title: Amendment
        author_id:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Author Id
      type: object
      required:
      - amendment
      title: AmendmentIn
...
    EncounterIn:
      properties:
        encounter_type:
          type: string
          pattern: ^(OUTPATIENT|INPATIENT|ED|SURGERY|TELEHEALTH|HOME)$
          title: Encounter Type
        department_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Department Id
        provider_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Id
        start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Time
        visit_number:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Visit Number
        reason:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Reason
      type: object
      required:
      - encounter_type
      title: EncounterIn
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...